+++ src/sensorthings_utils/frosty/sanitization.py
@@ -18,9 +18,10 @@ def to_odata_datetime(value: datetime | str) -> str:
         return value.isoformat()
     return str(value)
 
+#TODO: clarify exactly what this function sanitizes (i.e. the entities passed to it)
 def sanitize_get_request(
     root_url: str,
-    version: str | float | int | FrostVersions,
+    version: str | float | int,
     first_entity: str | SensorThingsEntityGroups | SensorThingsEntity,
     params: Optional[Mapping[str | FrostParams, Any]] = None,
     *,
@@ -30,12 +31,10 @@ def sanitize_get_request(
     """Normalize GET request URL and parameter keys for FROST queries."""
 
     entity_to_group = {v: k for k, v in ENTITY_GROUPS_TO_ENTITIES.items()}
-    normalized_root = root_url.rstrip("/")
     try:
         normalized_first_entity = SensorThingsEntityGroups(first_entity).value
     except ValueError:
         normalized_first_entity = entity_to_group[SensorThingsEntity(first_entity)].value
-    normalized_version = FrostVersions(str(version).lstrip("v")).value
     normalized_second_entity = ""
     normalized_first_entity_id = ""
 
@@ -52,12 +51,21 @@ def sanitize_get_request(
     normalized_params: dict[str, Any] | None = None
     if params:
         normalized_params = {FrostParams(key).value: value for key, value in params.items()}
-
+    
+    sanitized_root, sanitized_version = sanitize_root_url(root_url, version)
     url = (
-        f"{normalized_root}/v{normalized_version}/"
+        f"{sanitized_root}/v{sanitized_version}/"
         f"{normalized_first_entity}{normalized_first_entity_id}/{normalized_second_entity}"
     ).rstrip("/")
     return url, normalized_params
+
+def sanitize_root_url(
+        root_url:str,
+        version: str | int | float
+        )-> tuple[str, str]:
+    normalized_root = str(root_url.rstrip("/"))
+    normalized_version = FrostVersions(str(version).lstrip("v")).value
+    return (normalized_root, normalized_version)
 
 def merge_filter(params: dict[str, Any], extra_filter: str) -> dict[str, Any]:
     """Merge generated filter text with an existing $filter if present."""
